refactor: 문자열 하드코딩 열거형으로 관리
Github Issue: #52
